beehive-loader/requirements.txt
cassandra-driver
pika
